Split linter into two pages, DRY components
